First version of list of starred challenges of a user
