+++ src/app/pages/account/user-account/tab.model.ts
@@ -1,0 +1,4 @@
+export interface Tab {
+  name: string;
+  visible: boolean;
+}
