+++ src/app/pages/account/user-account/user-account-tabs.ts
@@ -1,0 +1,16 @@
+import { Tab } from './tab.model';
+
+export const TABS: { [key: string]: Tab } = {
+  overview: {
+    name: 'Overview',
+    visible: true,
+  },
+  challenges: {
+    name: 'Challenges',
+    visible: true,
+  },
+  starred: {
+    name: 'Starred',
+    visible: false,
+  },
+};
