+++ src/app/pages/account/user-account/user-profile-starred/user-profile-starred.component.html
@@ -1,1 +1,5 @@
-<p>user-profile-starred works!</p>
+<h2>Starred</h2>
+<mat-nav-list>
+  <a mat-list-item [routerLink]="['/', star.fullName]"
+    *ngFor="let star of stars">{{star.fullName}}</a>
+</mat-nav-list>
