+++ src/app/pages/account/user-account/user-profile-starred/user-profile-starred.component.ts
@@ -1,15 +1,33 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnDestroy, OnInit } from '@angular/core';
+import {
+  Challenge,
+  User,
+  UserService,
+} from '@sage-bionetworks/rocc-client-angular';
+import { Subscription } from 'rxjs';
+import { map } from 'rxjs/operators';
 
 @Component({
   selector: 'rocc-user-profile-starred',
   templateUrl: './user-profile-starred.component.html',
-  styleUrls: ['./user-profile-starred.component.scss']
+  styleUrls: ['./user-profile-starred.component.scss'],
 })
-export class UserProfileStarredComponent implements OnInit {
+export class UserProfileStarredComponent implements OnInit, OnDestroy {
+  @Input() user!: User;
+  stars: Challenge[] = [];
+  private subscriptions: Subscription[] = [];
 
-  constructor() { }
+  constructor(private userService: UserService) {}
 
   ngOnInit(): void {
+    const starsSub = this.userService
+      .listUserStarredChallenges(this.user.id)
+      .pipe(map((page) => page.challenges))
+      .subscribe((stars) => (this.stars = stars));
+    this.subscriptions.push(starsSub);
   }
 
+  ngOnDestroy(): void {
+    this.subscriptions.forEach((sub) => sub.unsubscribe());
+  }
 }
